refactor: model files name change
