+++ src/models/taskChecklistItemModel.ts
