+++ src/service/projectManagement/sendNotification/taskReminderNotification/notification.service.ts
@@ -1,6 +1,6 @@
 import userInfo from "../../../../models/userInfoModel";
 import Group from "../../../../models/groupModel";
-import Project from "../../../../models/projectModel";
+import Project from "../../../../models/groupProjectModel";
 import { RESPONSE_CODE } from "../../../utils/constants";
 import { AUTH_EMAIL } from "../../../../config";
 import { sendNotificationService } from "../sendNotification.service";
