+++ src/service/projectManagement/task/updateTask/updateTask.service.ts
@@ -1,4 +1,4 @@
-import Task from "../../../../models/taskModel";
+import Task from "../../../../models/projectTaskModel";
 import { RESPONSE_CODE } from "../../../utils/constants";
 
 export async function updateTaskService(req: any, res: any, next: any) {
